third_party/blink/web_tests/platform/mac-mac12-arm64/external/wpt/css/cssom/getComputedStyle-width-scroll.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] getComputedStyle() round-trips in presence of scrollbars
  assert_equals: width round-trips expected "85px" but got "70px"
Harness: the test ran to completion.
